feat: remove boot-screen gimmicks, split personal page into quniv/
- delete TerminalBoot popup, its boot.js data, and the GameOfLife RPG-style
  growth-cycle widget (used on both the dead AboutMe page and the live
  Team page's 'How We Work' section)
- delete HangingScroll and MotivationSection decorative components (only
  used by AboutMe)
- strip emoji flourishes from AboutMe bio/footer
- replace the fake typewriter terminal boot animation on News and Blog
  with a plain static placeholder
- move AboutMe.svelte, CV.svelte, and cvData.js into src/quniv/ — these
  were dead code (never routed from App.svelte) and are being staged
  for a future quniv.qtlab.dev personal subdomain; News.svelte stays
  under qtlab.dev
