--- src/data/boot.js
@@ -1,7 +1,0 @@
-export const bootSequence = [
-  '[qtlab.dev] Initializing...',
-  '[qtlab.dev] Systems online.',
-  '[qtlab.dev] Ready to build.',
-  'ACCESS GRANTED — Welcome, visitor.',
-  '> _ <',
-];
